+++ mpc_energy/CHANGELOG.md
@@ -5,7 +5,7 @@
 * Increased MPC grid import penalty and decreased EOD charge reward to reduce unnecessary grid charging 
 * Fixed issue where inconsistent Sigenergy load data caused an error
 * Fixed bad entity data retrival causing system exit
-
+* Added detection and improved logging if Sigenergy system goes offline
 
 ## 0.3.0
 * Added manual control option with price change trigger